alacritty-ligatures: add runtime dependencies
Ones of the changes from 0.5.0 is depending on the system fontconfig,
and one of the ligature changes forces the use of the system freetype.

One of the Rust libraries depends on the C++ stdlib.

At some point the build process stopped bundling the OpenGL libraries,
so add those back in too.